Move StoryPart to separate file
